Simpified it more
I reduced the functions down to one function named axisMove(); and it has parameters for the direct pin, digital output value, step pin, and the amount of steps. Based on the axisMove(); function I created a xMove(); and yMove(); function with parameters for the digital outputs and I am using them in a function called sixPasses(); which does what the name implies, the machine makes six passes. I may or may not need six passes in production, I might be able to reduce it to three especially when I get a diamond blade for it.

The next step will be to create a curvedCuts(); function that will be called after sixPasses(); and it will be the final cutting pass which will be slightly curved cuts on the pen blanks to give them a nice curvature and make the sanding much faster.

This code is way better than my previous code and it works great so far, however there's probably a much better way to do this and I look forward to finding out!

Live long and prosper,
Joseph Freedman